pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepersplunk"
version = "0.1.0"
description = "Adversarial SOC triage agent for Splunk, exposed as an MCP server for Claude."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Your Name" }]
keywords = ["mcp", "claude", "splunk", "soc", "siem", "triage", "security"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "Topic :: Security",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp>=1.2.0",
    "pydantic>=2.6",
]

[project.optional-dependencies]
splunk = ["splunk-sdk>=2.0.0"]
dev = [
    "pytest>=8.0",
    "python-dotenv>=1.0",
    "ruff>=0.3",
]

[project.scripts]
deepersplunk = "deepersplunk.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
